Cambio de ordenamiento en la lista de componente categoria, correccion del bug al eliminar un elemento de marcas

+++ app/Livewire/Forms/BrandForm.php
@@ -78,7 +78,9 @@ class BrandForm extends Form
 
     public function delete()
     {
-        Storage::disk('public')->delete($this->brand->logo);
+        if ($this->brand->logo) {
+            Storage::disk('public')->delete($this->brand->logo);
+        }
 
         $this->brand->delete();
